Introduce a BoltDB based Database Handler
This adds a new LDAPHandler for retrieving and storing the LDAP tree in
a BoltDB based database backend. Currently this Handler only supports
search (filtering is still handled in the frontend). It's meant as a
starting point to later add write capabilities to libregraph/idm.

Also this adds the new subcommand `idm boltdb load` to load the initial
dataset from an LDIF formatted file.