Fix error behaviour when receiving unsupported operation
Instead of returning an LDAPAdd Response with OperationsError on unknown
or unimplemented LDAP operations we now handle these conditions more
specifically.

If a valid Operation ID is received but we do not implement that
operation we return the correct Response of that Operation with the
error set to OperationsError. (This is currently the case for
LDAPModifyDN and LDAPCompoare operations).

If the OperationID does not match any of the LDAP Operation Request Tags
we just close the connection without returning an error. As basically
those packages are invalid LDAP client requests.